requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiomysql==0.2.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
click==8.2.0
coredis==4.22.0
Deprecated==1.2.18
Flask==3.1.1
flask-paginate==2024.4.12
frozenlist==1.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
orjson==3.10.18
packaging==25.0
pillow==11.2.1
priority==2.0.0
propcache==0.3.1
pybase64==1.4.1
pydantic==2.11.4
pydantic_core==2.33.2
pyhumps==3.8.0
Pympler==1.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
Quart==0.20.0
quart-auth==0.11.0
quart-flask-patch==0.3.0
quart-rate-limiter==0.11.0
quart-schema==0.22.0
quart-uploads==0.0.4
quart-wtforms==1.0.3
six==1.17.0
tqdm==4.67.1
types-aiofiles==24.1.0.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
yarl==1.20.0
msgpack
zlib_ng

# timm # will install the necessary torch packages for image tagging